requirements.txt
streamlit
pdfplumber
pandas
openpyxl
python-docx